Entering and exiting

+++ client/src/main/java/il/cshaifasweng/OCSFMediatorExample/client/CompliantsListController.java
@@ -38,13 +38,13 @@ public class CompliantsListController {
     private TableView<Complaint> complaints;
 
     @FXML
-    private TableColumn<Complaint, Integer> complaintId;
+    private TableColumn<Complaint, Long> complaintId;
 
     @FXML
-    private TableColumn<Complaint, Integer> userId;
+    private TableColumn<Complaint, Long> userId;
 
     @FXML
-    private TableColumn<Complaint, Integer> pLotIdCol;
+    private TableColumn<Complaint, Long> pLotIdCol;
 
     @FXML
     private TableColumn<Complaint, String> complaintDescription;
@@ -87,11 +87,11 @@ public class CompliantsListController {
 
     private Complaint selected;
 
-    private int getComplaintId;
+    private long getComplaintId;
 
-    private int getUserId;
+    private long getUserId;
 
-    private int getParkingLotId;
+    private long getParkingLotId;
     private boolean isRemoved;
